requirements.txt
pymupdf
rapidocr-onnxruntime
opencv-python
pillow
numpy
tqdm